REFACTOR: toast code to single file
