+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import {  Inter } from "next/font/google";
 import "./globals.css";
 
-import { ToastProvider } from "@/features/commons/ui/toast/context/toastProvider";
+import { ToastProvider } from "@/common/ui/toast";
 
 const inter = Inter({
   subsets: ['latin'],
