--- src/features/commons/ui/toast/context/useToast.tsx
@@ -1,10 +1,0 @@
-import { useContext } from 'react';
-import { ToastContext } from './toastProvider';
-
-export function useToast() {
-  const context = useContext(ToastContext);
-  if (!context) {
-    throw new Error('useToast must be used within a ToastProvider');
-  }
-  return context;
-}
